requirements.txt
numpy
pillow
opencv
xmltodict

